# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Family Budget Service API
  version: 1.0.0
  description: |
    Целевой контракт `/api/v1` для Android-клиента (spec 005, решения A-01…A-09).

    Код и спецификация совпадают: `tests/integration/openapi_coverage_test.go` проверяет
    обе стороны — и зарегистрированный роут без описания здесь, и описанную операцию без
    роута роняют `make test`.

    Соглашения:
    - Деньги — целые в минимальных единицах валюты семьи (`amount_minor`, int64). Валюта — у семьи (A-05).
    - Дата операции — календарная `YYYY-MM-DD` без времени; служебные метки — RFC3339 UTC (A-06).
    - `POST` для транзакций, бюджетов и категорий принимает клиентский `id` (любой валидный UUID);
      повтор с тем же `id` возвращает `200` с существующей записью, тело при этом игнорируется (A-07).
    - Дробные числа в двух единицах: доли 0…1 — `share`, `*_delta` и `utilization` в
      `StatsSummary.budgets[]` (`BudgetProgress`); проценты 0…100 — `percentage` и `utilization`
      в `Budget` (`GET /api/v1/budgets`).
    - Списки: `limit` (по умолчанию 50, максимум 200) и `offset`; `meta.pagination.total` — общее число (A-08).
    - Ошибка всегда одной формы: `{"error": {...}, "meta": {...}}` (A-08).

    Коды ошибок: `INVALID_REQUEST` / `INVALID_ID` (400 — тело или id не разобрались),
    `REQUEST_TIMEOUT` (408 — тело не успело загрузиться), `PAYLOAD_TOO_LARGE` (413),
    `UNAUTHORIZED` (401), `FORBIDDEN` (403), `NOT_FOUND` / `<ENTITY>_NOT_FOUND` (404),
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Выдать bearer-токен
      description: |
        Публичный роут. Токен отдаётся один раз и хранится на сервере только как SHA-256 (A-01).
        Ответ на неизвестный email и на неверный пароль одинаков — `401 INVALID_CREDENTIALS`.
        Лимитер: 10 попыток / 5 минут на IP и 20 / час на email (A-09).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Токен выдан
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/LoginResponse' }
                  meta: { $ref: '#/components/schemas/Meta' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/InvalidCredentials' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Отозвать текущую сессию
      description: Доступно любой роли. Токен из заголовка становится недействительным немедленно.
      responses:
        '204': { description: Сессия отозвана }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/sessions:
    get:
      operationId: listSessions
      tags: [auth]
      summary: Активные сессии текущего пользователя
      description: Доступно любой роли; чужие сессии не видны даже админу.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Список сессий
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Session' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/sessions/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    delete:
      operationId: revokeSession
      tags: [auth]
      summary: Отозвать сессию
      description: Только собственная сессия текущего пользователя; чужая — `404`.
      responses:
        '204': { description: Сессия отозвана }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/me:
    get:
      operationId: getCurrentUser
      tags: [me]
      summary: Текущий пользователь
      description: Доступно любой роли.
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateCurrentUser
      tags: [me]
      summary: Изменить имя и email
      description: Доступно любой роли. Роль и `is_active` через этот роут не меняются.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/me/password:
    put:
      operationId: changePassword
      tags: [me]
      summary: Сменить свой пароль
      description: |
        Доступно любой роли. Успешная смена отзывает все сессии пользователя, кроме текущей (A-01).
        Неверный `current_password` — `401 INVALID_CREDENTIALS`; сессия при этом жива,
        разлогинивать клиента не нужно.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ChangePasswordRequest' }
      responses:
        '204': { description: Пароль изменён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: |
            `UNAUTHORIZED` — токен отсутствует, просрочен или отозван;
            `INVALID_CREDENTIALS` — неверный `current_password`. Различать по `error.code`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/family:
    get:
      operationId: getFamily
      tags: [family]
      summary: Данные семьи
      description: Доступно любой роли.
      responses:
        '200': { $ref: '#/components/responses/FamilyOk' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      operationId: updateFamily
      tags: [family]
      summary: Изменить семью
      description: |
        Только admin. Смена `currency` при наличии транзакций или сверок запрещена — `409 CURRENCY_LOCKED` (A-05).
        `timezone` — IANA-имя, в нём считаются «сегодня» и границы периодов (A-06).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateFamilyRequest' }
      responses:
        '200': { $ref: '#/components/responses/FamilyOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: Пользователи семьи
      description: Только admin.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      operationId: createUser
      tags: [users]
      summary: Создать пользователя
      description: |
        Только admin. Инвайтов нет: админ задаёт пароль и сообщает его лично,
        пользователь меняет его через `PUT /api/v1/me/password` (A-03).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/users/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    get:
      operationId: getUser
      tags: [users]
      summary: Пользователь по id
      description: Только admin.
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateUser
      tags: [users]
      summary: Изменить имя и email пользователя
      description: |
        Только admin. Email, занятый другим пользователем, — `409 EMAIL_TAKEN`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    patch:
      operationId: patchUser
      tags: [users]
      summary: Изменить роль или активность
      description: |
        Только admin. Деактивация (`is_active: false`) отзывает все сессии пользователя (A-04).
        Понижение роли или деактивация последнего активного админа — `409 LAST_ADMIN`;
        деактивация себя — `409 CANNOT_DEACTIVATE_SELF`. Оба поля применяются одной
        записью: при `409` не изменено ни одно из них.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchUserRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/users/{id}/password:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    put:
      operationId: setUserPassword
      tags: [users]
      summary: Задать пароль пользователю
      description: Только admin. Текущий пароль не требуется; все сессии пользователя отзываются.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetPasswordRequest' }
      responses:
        '204': { description: Пароль задан }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Категории семьи
      description: admin и member.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: type
          in: query
          required: false
          schema: { $ref: '#/components/schemas/CategoryType' }
      responses:
        '200':
          description: Список категорий
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Category' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      operationId: createCategory
      tags: [categories]
      summary: Создать категорию
      description: admin и member. Повтор с тем же `id` возвращает `200` с существующей
        категорией; остальное тело при этом игнорируется (A-07).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCategoryRequest' }
      responses:
        '201': { $ref: '#/components/responses/CategoryOk' }
        '200': { $ref: '#/components/responses/CategoryOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/categories/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    get:
      operationId: getCategory
      tags: [categories]
      summary: Категория по id
      description: |
        admin и member. Удалённая (`is_active: false`) категория по-прежнему отдаётся `200` —
        по ней разрешается имя категории у сохранённых транзакций; из списка она исчезает.
      responses:
        '200': { $ref: '#/components/responses/CategoryOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Изменить категорию
      description: |
        admin и member. Тип категории не меняется — у неё уже есть транзакции.
        Удалённая (`is_active: false`) категория не редактируется — `404`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCategoryRequest' }
      responses:
        '200': { $ref: '#/components/responses/CategoryOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Удалить категорию
      description: |
        Только admin. Категория помечается неактивной (`is_active: false`) и перестаёт выводиться
        в списке, ответ `204`; её транзакции сохраняются. Подкатегории деактивируются вместе
        с родителем. Повторный `DELETE` уже неактивной категории — `404`.
      responses:
        '204': { description: Категория удалена }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: Счета семьи
      description: admin и member. Порядок — по имени без учёта регистра.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: archived
          in: query
          required: false
          description: '`true` добавляет в список архивные счета'
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Список счетов
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Создать счёт
      description: |
        admin и member. Повтор с тем же `id` возвращает `200` с существующим счётом; остальное тело
        при этом игнорируется (A-07). Имя уникально в семье без учёта регистра и пробелов по краям,
        архивный счёт его тоже занимает — `409 ACCOUNT_NAME_EXISTS`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountRequest' }
      responses:
        '201': { $ref: '#/components/responses/AccountOk' }
        '200': { $ref: '#/components/responses/AccountOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/accounts/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Переименовать, архивировать или вернуть счёт
      description: admin и member. Архивный счёт скрыт из списка без `?archived=true` и не назначается
        новым операциям.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAccountRequest' }
      responses:
        '200': { $ref: '#/components/responses/AccountOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Удалить счёт
      description: |
        Только admin. Счёт с операциями или сверками не удаляется — `409 ACCOUNT_IN_USE`;
        перевыпущенную карту архивируют.
      responses:
        '204': { description: Счёт удалён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/v1/holdings/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    put:
      operationId: updateHolding
      tags: [holdings]
      summary: Переименовать, сменить вид, архивировать или вернуть позицию
      description: |
        admin и member. `side` не меняется: поля в запросе нет, присланное игнорируется. Архив только
        прячет позицию из списка без `?archived=true` — в капитале она остаётся по своим снимкам;
        прекращение стоимости — снимок `0`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateHoldingRequest' }
      responses:
        '200': { $ref: '#/components/responses/HoldingOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteHolding
      tags: [holdings]
      summary: Удалить позицию
      description: |
        Только admin. Снимки удаляются вместе с позицией, и прошлые значения капитала меняются — это
        исправление ошибочно заведённой позиции. Проданное имущество — снимок `0` и архив.
      responses:
        '204': { description: Позиция удалена }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/v1/backups/{name}:
    parameters:
      - { $ref: '#/components/parameters/BackupName' }
    delete:
      operationId: deleteBackup
      tags: [backups]
      summary: Удалить бэкап
      description: Только admin.
      responses:
        '204': { description: Файл удалён }
        '400':
          description: Имя не соответствует шаблону `backup_YYYYMMDD_HHMMSSmmm.db`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Непрозрачный токен из `POST /api/v1/auth/login` (A-01): 32 байта `crypto/rand` в base64url.
        Не JWT — сервер хранит SHA-256 и проверяет срок и активность пользователя на каждом запросе.
        Скользящий срок — 30 дней без активности, потолок — 180 дней от выдачи.
...
  schemas:
...
    CategoryType:
      type: string
      enum: [income, expense]
...